+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 [//]: # (current developments)
 
+* Use `compression.zstd` or `backports.zstd` (before / after Python 3.14), instead of
+  `zstandard / python-zstandard. (#137)
+* Set minimum Python version to 3.10 (#138)
+
 ## 0.12.0 (2025-06)
 
 * Skip setting permissions if `tarinfo.mode` is `None`. (#140)